requirements.txt
mysql-connector-python
pandas
numpy
datetime
schedule
time
random
uuid
logging
multiprocessing
multithreading